Use only a public key for decryption
The use of a tar file was admittedly anti-pattern, a sign of my own
ignorance concerning all things encrypted. It makes much more sense
to generate the surrounding metadata from the private key as
necessary, which we now do. The former RAPTOR_GPG parameter has
become RAPTOR_KEY and should simply be an authorized GPG key which
can be used without a password. The documentation has been updated
to this effect.

The docs also include what tags are sent to the server for added
clarity.